move getAddressType and getByteCount methods to common helpers
